fertilization changes

+++ app/dto/FertilizerOperationDto.java
@@ -12,8 +12,8 @@ public class FertilizerOperationDto {
 	public Date endTo;
 	public Double quantity;
 
-	public static int sumOfQuantity(List<FertilizerOperationDto> list) {
-		int sum = 0;
+	public static double sumOfQuantity(List<FertilizerOperationDto> list) {
+		double sum = 0;
 		for (FertilizerOperationDto fo : list) {
 			sum += fo.quantity;
 		}
